Principal-main/desafio3/web/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.7
redis==5.2.0
